moving around functions to make easier to test
